+++ System.IO.Ports/Properties/AssemblyInfo.cs
@@ -5,14 +5,14 @@ using System.Runtime.InteropServices;
 // General Information about an assembly is controlled through the following 
 // set of attributes. Change these attribute values to modify the information
 // associated with an assembly.
-[assembly: AssemblyTitle("System.IO.Ports")]
+[assembly: AssemblyTitle("nanoFramework.System.IO.Ports")]
 [assembly: AssemblyCompany("nanoFramework Contributors")]
-[assembly: AssemblyProduct("System.IO.Ports")]
+[assembly: AssemblyProduct("nanoFramework.System.IO.Ports")]
 [assembly: AssemblyCopyright("Copyright (c) .NET Foundation and Contributors")]
 
 ////////////////////////////////////////////////////////////////
 // update this whenever the native assembly signature changes //
-[assembly: AssemblyNativeVersion("100.1.1.3")]
+[assembly: AssemblyNativeVersion("100.1.2.0")]
 ////////////////////////////////////////////////////////////////
 
 // Setting ComVisible to false makes the types in this assembly not visible